BE:refactor(chats/messages): update error handling

+++ backend/routers/messages.py
@@ -17,7 +17,12 @@ async def get_messages_by_chat(
 ):
     msg_repo = MessageRepository(session)
     chat_repo = ChatRepository(session)
-    chat_repo.is_chat_member(chat_id, current_user.id)
+
+    chat = await chat_repo.get_chat_by_user_id(chat_id, current_user.id)
+    if not chat:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Chat not found"
+        )
 
     return await msg_repo.get_messages_by_chat_id(chat_id)
 
@@ -36,16 +41,22 @@ async def add_message(
     msg_repo = MessageRepository(session)
     chat_repo = ChatRepository(session)
 
-    await chat_repo.is_chat_member(chat_id, current_user.id)
+    chat = await chat_repo.get_chat_by_user_id(chat_id, current_user.id)
+    if not chat:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Chat not found"
+        )
 
     message = await msg_repo.add_new_message(msg_data.content, chat_id, current_user.id)
     broadcast_payload = {"type": "new_message", "payload": message.model_dump()}
-    await manager.broadcast_to_chat(json.dumps(broadcast_payload, default=str), chat_id, session)
+    await manager.broadcast_to_chat(
+        json.dumps(broadcast_payload, default=str), chat_id, session
+    )
 
     return message
 
 
-@router.put(
+@router.patch(
     "/{msg_id}",
     response_model=MessageRead,
 )
@@ -56,12 +67,12 @@ async def edit_message(
     session: SessionDependency,
 ):
     msg_repo = MessageRepository(session)
-    chat_repo = ChatRepository(session)
 
-    message = await msg_repo.get_message_by_id(msg_id)
-
-    await chat_repo.is_chat_member(message.chat_id, current_user.id)
-    msg_repo.is_user_message_sender(message.sender_id, current_user.id)
+    message = await msg_repo.get_message_by_user_id(msg_id, current_user.id)
+    if not message:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Message not found"
+        )
 
     message.content = msg_data.content
     updated_message = await msg_repo.update_message(message)
@@ -69,7 +80,9 @@ async def edit_message(
         "type": "edit_message",
         "payload": updated_message.model_dump(),
     }
-    await manager.broadcast_to_chat(json.dumps(broadcast_payload, default=str), message.chat_id, session)
+    await manager.broadcast_to_chat(
+        json.dumps(broadcast_payload, default=str), updated_message.chat_id, session
+    )
 
     return updated_message
 
@@ -81,16 +94,19 @@ async def delete_message(
     session: SessionDependency,
 ):
     msg_repo = MessageRepository(session)
-    chat_repo = ChatRepository(session)
 
-    message = await msg_repo.get_message_by_id(msg_id)
-
-    await chat_repo.is_chat_member(message.chat_id, current_user.id)
-    msg_repo.is_user_message_sender(message.sender_id, current_user.id)
+    message = await msg_repo.get_message_by_user_id(msg_id, current_user.id)
+    if not message:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Message not found"
+        )
 
     await msg_repo.delete_message(message)
+
     broadcast_payload = {
         "type": "delete_message",
         "payload": {"id": msg_id, "chat_id": message.chat_id},
     }
-    await manager.broadcast_to_chat(json.dumps(broadcast_payload, default=str), message.chat_id, session)
+    await manager.broadcast_to_chat(
+        json.dumps(broadcast_payload, default=str), message.chat_id, session
+    )
